enable cleartype on canvas elements
DRQS 32281407
Previous logic disables cleartype if we are drawing to an image buffer (because
cleartype would not work in compositing mode).  We'll just enable this for now
since we want cleartype on canvas elements.  In a future change, we'll fix this
to be based on the -webkit-font-smoothing setting.

This patch will not apply with the latest webkit trunk.  In order to apply this
when we update, we should make this change in PlatformContextSkia.cpp:
PlatformContextSkia::couldUseLCDRenderedText()