use fontSmoothing setting from fontDescription
DRQS 32281407
This change is a better fix that would replace the change introduced in the
bugfix/fontsmoothing branch.  I will undo the change from that branch after
this has landed, then delete that branch so that we don't apply it to
subsequent rebases with upstream.

In this change, we use the fontSmoothing setting from the FontDescription to
determine the lfQuality whenever we call paintSkiaText.  If the fontSmoothing
is AutoSmoothing, then we will use the lfQuality from the FontPlatformData.

The benefit of this change over the bugfix/fontsmoothing change is that:

    * this is a simpler code change

    * it will be easier to contribute back because it only affects Skia and not
      every platform (upstream: https://bugs.webkit.org/show_bug.cgi?id=86776)

    * it doesn't create a separate FontPlatformData for each fontSmoothing
      setting.  Skia has its own font cache, so we don't need to use the one
      from platform/graphics/FontCache.cpp