make list marker have same size as first text
DRQS 32534047
updateMarkerLocation() is called whenever a relayout happens.  If we detect
that the marker's font and the first text's font are different, then we make
sure the marker gets the same font as the first text.

As a hack, right now we force a style recalc by virtue of removing the marker
and readding it back to the parent.