Implement multicolumn header spanning
DRQS 33399541
Note that this only supports a single header per multicolumn div, and the
header should be at the beginning of the div.

When the header div's layout is complete, we store the header's height and
column span in the ColumnInfo structure.

Then, when layout of the remaining content is done, the vertical position of
the first linebox in a column that contains the spanning header is adjusted by
the spanning header's height, in addition to the remaining logical height that
was used to push the linebox onto the next column.

The change in RenderBox is what causes the header's width to span across
multiple columns, instead of using the availableLogicalWidth of a single
column.

The change in paintColumnContents is to expand the clip rect when painting the
first column, only if it contains a spanning header.