Don't include pagination strut in selection
DRQS 33399541
When selectionTop() and selectionBottom() use the previous and next root boxes
respectively, make sure we don't include the pagination strut that we
introduced due to the spanning header.

This fixes the issue where the selection highlight of the lines directly below
the spanning header extends over the spanning header.