Fix clipping and hit testing issues
DRQS 33399541
Add 'adjustColRectForSpanningHeader' function, which adjusts the rect obtained
from columnRectAt to take into account the spanning header.  This includes
expanding the rect for the first column, and for subsequent columns within the
column span, move the rect below the spanning header.  This is a refactor of
the adjustment we did previously in paintColumnContents.

This function is used in other places to fix hit-testing (which allows us to
select items within the spanning header) and repaint rect calculations (which
was causing some weird paint behavior, particularly with selection highlights).