Disable the V8RecursionScope assertion
DRQS 34552506
This assertion makes it impossible to allow non-webkit related JavaScript run
in the same isolate as WebKit, without adding a MicrotaskSuppression guard,
which is impossible without a bunch of levelization hacks.

The only purpose of this assertion is to ensure that if the JavaScript modifies
the DOM, then the mutation events will be fired once the V8RecursionScope goes
out of scope.

In our case, we would construct a V8RecursionScope around an entire block of
code that manipulates the DOM, not at each function call site.  Therefore it
would be the responsibility of higher level code to ensure that they add
V8RecursionScope in the right places to cause mutation observer to do its work.