Don't add BR at start of paragraph
DRQS 36813987
WebKit adds a BR element when it needs to split paragraphs that were
potentially joined after deleting/pruning nodes for a move.  However, if we are
already at the start of a paragraph, then there is no need to add a BR element.

This fixes an issue where tables moved down (new BRs added before the TABLE
element) whenever the cell contents were listified and unlistified.