Ensure no selection gaps on left of list markers
Since we are about to reintroduce selection gaps (I'm about to commit a proper
fix for bugfix/selectionGapTransform), we need to ensure that we do not fill
selection gaps around list markers.

There's probably a better way to do this.  Right now, we just adjust
logicalLeftSelectionOffset to not include the padding list elements have for
the list markers.

Also, in order to prevent selection gaps from getting filled between list
markers and the first box under the list item, we just set the
isPreviousBoxSelected flag to false if it was a list marker.