Fill left/right selection gaps only for tablecells
DRQS 37611655
Modify RenderBlock::getSelectionGapInfo so that left and right gaps are only
filled for table cells.  Also, allow table cells to be selection leafs if they
have no children, otherwise they will not highlight.

Since not all blocks fill left and right selection gaps, we need to change the
way lastLogicalLeft and lastLogicalRight are calculated.  If we are filling the
selection gap, then we use the old logic, which tends to put us against the
border of the root block.  Otherwise, we use the logical left/right of the last
selected line.

Additionally, if the last selected line is where selection started, then we
need to push lastLogicalLeft to where the first selected box's selection rect
begins.