Change the way list layout is done
DRQS 37467409
In upstream WebKit, each OL/UL element is given 40px padding, and the list
marker's margin is adjusted to position the marker to the logical left of the
LI block.  This is irrespective of the font size of the list marker, and can
result in the list marker being positioned off the screen.

We change this such that 40px padding is not applied automatically, instead the
LI block is given an additional marginStart that is equivalent to the width of
the marker.

The following changes are made:
* The -webkit-padding-start for lists is removed.
* Added RenderBlock::additionalMarginStart, which is overridden in
  RenderListItem to return the width of the list marker.  It is also overridden
  in RenderBlock so that block children of list elements return the marginStart
  of the previous list item.  This is how nested lists get their margin.
* RenderListMarker does not include a space in its suffix.
* The minPreferredLogicalWidth of RenderListMarker includes the marker padding.