Remove special-case of tables for special elements
DRQS 37576223
The firstInSpecialElement and lastInSpecialElement functions handle tables
specially, in a way that doesn't make much sense.  I traced this behavior back
to revision 12358 in WebKit's SVN, which was committed in 2006.  It is likely
that this special case for tables made sense back then, but no longer makes
sense due to changes in other parts of the code since then.

The original behavior was causing DeleteSelectionCommand::initializeStartEnd to
modify the selection start/end positions to be outside the table element, which
resulted in the entire table getting deleted.