Fix computeBlockPreferredLogicalWidths
DRQS 38468073
The additionalMarginStart was not being taken into account when computing the
preferred logical widths of block children.  We treat the additionalMarginStart
as a fixed margin, so it is always added to marginStart.

Also, fix RenderListItem::additionalMarginStart to updateMarkerLocation before
calculating its minPreferredLogicalWidth.  This is necessary to ensure that the
marker has been parented correctly, otherwise minPreferredLogicalWidth will not
return the right value.  Is the const_cast evil?  Well, it is also used
upstream in RenderBox::minPreferredLogicalWidth, so I suppose it is fine for
this project.