Fix lastLogicalLeft when selectionStartsOnLastLine
Previously, if the last selected line had no selection rect, our
lastLogicalLeft would be flushed against the left of the root block, because
the returned selectionRectForRepaint would be empty.

Now, when selection starts on the last line, we initialize lastLogicalLeft to
the *right* of the line (this is where the line ending gap begins).  Then, if
the line has a selection rect, we move the lastLogicalLeft to the selection
rect's left.