Allow line ending highlight to go into padding
This is necessary for right-aligned text.  If there is no padding, then the
line ending will not be highlighted.  However, if there is padding from the
line to the root block, then we can make use of that space for the highlight.
While it is technically bad to paint into the padding area, this may be an
exceptional case.

Also, adjust the width of the line ending gap to be based on the line height,
instead of being hardcoded to five pixels.  Now, the width will be one quarter
the line height.