Fix gap highlighting around list markers
The previous commit "Ensure no selection gaps on left of list markers" was
written based on the previous list layout algorithm where the list element was
given a padding of 40px.

Now since this is no longer the case, logicalLeftSelectionOffset needs to use
additionalMarginStart on the list item instead of paddingStart on the list
element.

Also, this commit fixes logicalRightSelectionOffset too so that lists with RTL
direction don't have gap highlights around the list markers.