Refactor ApplyBlockElementCommand::doApply
There is some useful functionality here, and we want to use it in the
BlockCommand class (and it can be used elsewhere too).

The logic before calling formatSelection has been moved to
CompositeEditCommand::prepareForBlockCommand.  This essentially does some
checks, then captures the current selection state so that it can be restored
after formatSelection, in the new CompositeEditCommand::finishBlockCommand
method.