Add blockExtentStart and blockExtentEnd
This is a greedy block extend algorithm.  If it is at a block, it walks up to
the highest parent that has no rendered siblings before or after it.  If it has
a sibling, it keeps eating it up until it comes to a block, otherwise it just
keeps moving up to the root.