Skip check in worldForEnteredContextWithoutContextCheck
We can't check the context in this case because the context may not be fully
initialized yet.  Note that this may break if we are not in a webkit context
when this is called, but it's not clear how else we can avoid this.  So far it
seems that this is called from within a webkit context, but we will have to see
once this is integrated with blp.