Expose ShadowingAtExitManager
Make ShadowingAtExitManager available outside unit tests.  We intend to use
this when the BrowserMain is running in a secondary thread.  In order to ensure
that objects are destructed in the BrowserMain thread instead of the
application's main thread (upstream chromium code depends on this), we will
create a ShadowingAtExitManager in the BrowserMain thread and join that thread
before the AtExitManager in the main thread is destroyed.