Refactor MessageLoop's Run
This is to facilitate integration of chromium's run loop with the application's
main run loop.  Chromium's actual run loop happens in DoRunLoop.  The Run
method does a bunch of stuff before eventually getting to DoRunLoop, and also
does a bunch of stuff after returning DoRunLoop.

This change refactors Run a bit to make it possible to do all that "before"
work and "after" work independently of calling DoRunLoop.  This gives us a
little more flexibility in how the run loop is used.  For example, we can
simply do the "before" work, then return back to the application to do its own
run loop, then do the "after" work during shutdown.