Add flag to disable hooks on root window
By default, Chromium hooks into the root window's WNDPROC to get the
WM_WINDOWPOSCHANGED and WM_SETTINGCHANGE messages.  However, the hooking
mechanism (in hwnd_subclass.cc) is not threadsafe, so things break randomly if
the root window does not belong to the browser-main thread.

This change introduces a flag to disable hooking into the root window.  In this
case, the application will be responsible for notifying Chromium when it
receives these two messages.