Expose window.open flags
The third parameter for 'window.open' allows the script to specify a bunch of
properties for the new window, for example position and size.  However, these
properties are not propagated to the WebContentsDelegate::WebContentsCreated
method.

This change exposes these properties by defining a a 'ContentCreatedParams'
class, which gets populated with these properties and passed on to the
WebContentsDelegate::WebContentsCreated method.

We also expose some blpwtk-specific flags (these should not be upstreamed).