Expose 'bbClipboard' for clipboard manipulation
'bbClipboard' is exposed on the 'window' object and provides various
functions for manipulating the clipboard through javascript, such as
emptying the clipboard, getting and setting custom strings, quering if a
custom format is set on the clipboard and setting a custom array buffer.

Note the 'bb' prefix, which warns that this is something we have exposed
for Bloomberg specific needs, and will not be upstreaming as is.