Add the ability to build sandbox in a DLL module
TREQ 2526677

This change takes advantage of the existing 'SANDBOX_EXPORTS' behavior, with
the additional feature that the sandbox code can be linked into a DLL instead
of linking directly into the executable.

When spawning the target process, the initial token will be preserved after the
process is launched (but still in suspended mode).  At this point, the DLL
containing the sandbox code will be injected using a remote thread running
inside the target process, created from the broker process.  After this, the
initial token is closed and execution resumes as normal.

Also, the exported symbols are now obtained from 'SandboxModule' instead of
'MainModule'.  If the DLL was not injected, then 'SandboxModule' is treated
equivalent to 'MainModule'.