Enable sandboxing for blpwtk2
We build sandbox.lib with 'SANDBOX_EXPORTS', and link it into blpwtk2.dll and
blpwtk2_subprocess.exe.  In order for this to work, we have to make
blpwtk2_subprocess.exe load blpwtk2.dll dynamically inside WinMain, because the
broker needs to load the exe image to obtain the exported offsets.  This was
causing significant delays if blpwtk2_subprocess.exe implicitly loads
blpwtk2.dll as well.

Also, we need to explicitly destroy the sandbox singletons when shutting down
ToolkitImpl, because if we wait for WinMain to return, this causes strange
behavior in the BrokerServiceBase destructor (this is elaborated further in the
sandboxDLL branch).