Auto-generate version files
The following files now get generated during the build:
* src/blpwtk2/public/blpwtk2_products.h
* src/blpwtk2/public/blpwtk2_version.h
* src/blpwtk2/public/blpwtk2_version.cc

The 'bb_version' gyp variable can be used to specify the version in the
following format:
    ##.#.####.##_bb##

The first component before '_' is the chromium version, and must match the
version directory next to the 'src' directory.  The second component is the
Bloomberg patch version.

The 'bb_version' value, if specified, is embedded in the product filenames.
Also, the V8 dll's product name is renamed to blpv8.<bb_version>.