Allow explicit destruction of singletons
Classes that implement SingletonBase (e.g. BrokerServicesBase) are
automatically deleted using Microsoft's '_onexit' CRT extension.

However, this causes issues when the sandbox is linked into a DLL, and the DLL
is implicitly loaded by the EXE.  In this case, the BrokerServicesBase object
is destroyed from the DLL CRT after WinMain has returned.

It seems like process handles are invalidated before that point, because the
TargetProcess destructor (which gets invoked from the job-thread) never returns
from waiting for the process to exit, even though the process has actually
exited.  It looks like the job-thread "disappears", because BrokerServiceBase
finishes waiting for it.  But it then crashes when calling FreeResources,
because it is trying to free a job that has been partially freed by the
job-thread (which just didn't get around to cleaning up after the free).

The simple solution to this problem is to expose a method that we can call
explicitly before returning from WinMain, which will cleanup all the
singletons.  This causes the BrokerServicesBase to wait for the job-thread to
complete successfully.