Add bbSetDataNoRelayout
DRQS 42070031

This is an extension for optimizing the case where we set text data, and know
that there is no relayout necessary.  Ideally, WebKit should detect this
situation automatically based on conditions like:
   * absolutely positioned
   * fixed size
   * contains a single text node (the one whose data is being set)
   * has "overflow: hidden"
   * has "white-space: nowrap"
   * has "text-align: left"

However, while we tweak this condition-set, we'll just use this separate entry
point to enable the optimized behavior.