Add render-process-affinity to WebContents
When the render-process-affinity property is set to anything other than
kNoProcessAffinity, then all web content will be rendered in that
render-process, regardless of whatever rules are used in upstream chromium to
manage render processes.

This allows us to do things like ensure two WebContents are always routed to
the same render-process.  Also, the WebContents will be routed to the same
render-process even when loading different URLs.

This change gives embedders completely manual control over the
creation/destruction of render-processes.  In order to create a render-process,
they can call RenderProcessHostImpl::GenerateUniqueId() then create a
WebContents with affinity to that unique id.