Allow in-process and out-of-process renderers
In upstream chromium, running an in-process renderer was done by specifying
--single-process on the command-line.  With this command-line argument, only
one RenderProcessHostImpl was created.

With this change, RenderProcessHostImpl now accepts an 'is_in_process'
argument.  This allows embedders to create in-process RenderProcessHost
explicitly, even without the --single-process command-line.

Combined with the ability to create WebContents with renderer affinity, this
change effectively allows embedders to create WebContents that run in-process
and also WebContents that run out-of-process, within the same application.

If no affinity is specified, then the RenderProcessHostImpl's is_in_process
property is initialized based on whether --single-process is specified, so
upstream applications like content_shell.exe continue to work as normal.