Don't let our pump starve in AUTOMATIC mode
When we are in AUTOMATIC mode, work is already done in a low priority WM_TIMER
event.  If there is a Windows message pending before we even start to do any
work, then it is likely that the application is firing WM_TIMER events of its
own.  In this case, we should allow at least one iteration of work to be done,
otherwise our pump will starve.

The check for pending messages at the beginning of doWorkUntilNextMessage()
should really only be used for the MANUAL pump mode, where we can expect the
application to call us back when it is done processing the message.