Make RWHV::OnPaint a no-op if we are hidden
The comment in RenderWidgetHostViewWin::WasHidden() says that we should ignore
paint messages when we are hidden.  However, the OnPaint() method does not do
this.

This was causing a DCHECK failure in RenderWidgetHostImpl::GetBackingStore(),
which checks that we are not painting when hidden.  It's still unclear why the
paint event is being dispatched when the window is hidden, but this error was
observed when trying to embed blpwtk2 inside the old blpwtk_shell using a
browser plugin, when using the RENDERER_MAIN thread-mode (i.e. the browser
thread is running separately from the main thread).

For now, just make this a no-op until we can investigate further why the paint
message is being dispatched.