Implement custom context menus
Summary:
Implement forwarding of the custom context menu action when an item is
being selected from the context menu. The outcome is that when a custom
context menu item is selected, the corresponding JavaScript binding
function is being called.

Test Plan:
Tested that the custom context menu items in the Inspector
work as expected.

Differential Revision: https://uiinf.phab.dev.bloomberg.com/D651