Add navigation controls
Add methods to WebView to go back and forward, and to stop loading the current
page.  There is also a new event added to WebViewDelegate to notify
applications about changes in the navigation state (whether back/forward can
actually be called).

Also, update the shell program to provide UI for these controls.