Add didFinishLoad and didFailLoad callbacks
WebViewImpl now observes the WebContents using the WebContentsObserver
interface.  Right now, we only listen for DidFinishLoad and DidFailLoad
callbacks, and invoke the corresponding methods on the WebViewDelegate, only if
these events are happening for the main frame.

The contract for mainFrame() has been changed such that the app needs to wait
for didFinishLoad() instead of didNavigateMainFramePostCommit().