Buffer WebViewProxy::move() calls
While we wait for the impl thread to process the move, we shouldn't schedule
more moves, otherwise there will be a backlog of moves when we drag the window
really quickly.

Instead, store the last move request inside the WebViewProxy, and when the impl
thread finishes processing the move, let it post an ack back to the proxy
thread.  When the proxy thread receives the ack, it will send the last move
request back to the impl thread (if it was different from what the impl thread
just processed).