Implement printing
Expose some new objects into the DOM to allow scripts to manipulate printing
behavior.  This is a temporary step until we can take advantage of CSS page
media when it is implemented upstream.

The new print utility in blpwtk2 uses this information from the DOM to perform
the actual printing for Win32.