Load devtools frontend from a pak file
Instead of loading devtools from the resources/inspector directory, load it
from a pak file, which is generated by a new blpwtk2 target.  The pak file
makes it easier to deploy the devtools on the machines that need it.

Also, make loadInspector undefined unless the devtools pak file was present at
startup.  A new 'hasDevTools' method has been added to Toolkit to let
applications know about this state.