Allow per-renderer in-process-plugins
Chromium has a "--in-process-plugins" flag to make plugins run in-process in
the renderer.  However, this makes *all* renderers run plugins in-process, and
therefore disables all sandboxing.

This change adds a flag to RenderProcessHost, which allows the embedder to
decide, for each renderer, whether in-process plugins should be used.  When
this flag is set for a renderer, sandboxing is disabled for that renderer, and
the "--in-process-plugins" switch is added to the renderer's command line.  If
the renderer is in-process, then we just force the flag into RenderProcessImpl
to avoid having to add "--in-process-plugins" to the browser's command line.