Add Toolkit API to use in-process plugins
This flag is set on a per-renderer basis, instead of a per-WebView basis,
because using in-process plugins has process-wide side effects (sandboxing is
disabled).

Part of this change also refactors ToolkitImpl's mapping from renderer to
hostId.  In this change, this mapping has been moved to the Statics component,
and we now have a mapping from renderer(id) to RendererInfo, which contains the
hostId and also a flag saying whether or not to use in-process plugins.  More
per-renderer configuration can be added here in the future.

When a RenderProcessHost is created, we query this map and use the recently
added SetUsesInProcessPlugins method if we want that renderer to use in-process
plugins.