Allow host to handle external protocol navigation (i.e. 'mailto:')
An external protocol is typically one that requires another program
to be launched or communicated to in order to handle it. A typical
example of this is 'mailto:' or in some case 'ftp:' URLs. Those can
also be custom, host specific protocols.

Provide a notification through the blpwtk2::WebViewDelegate interface
when such a navigation request is made. The callback function,
blpwtk2::WebViewDelegate::handleExternalProtocol is always called
on the application main thread.