Create in-process renderer only when it is used
Previously, when running in RENDERER_MAIN mode, the in-process renderer would
be created as soon as the BrowserThread is created.  With this change, the
in-process renderer is only created when a WebView using it is created the
first time.

Also cleanup BrowserThread to make the initial wait an implementation detail.