Fix patch for v28
The changes to WebCore.gypi and Document.idl failed to apply in the previous
merge because git's rename detection did not find them.  Also, because we moved
the code that starts the in-process renderer thread, we need to also apply the
relevant changes to that code.  Also, we don't need RenderProcessHostImpl's
GetInProcessRendererThreadForTesting method, and the corresponding global
variable, so clean that up as well.