Add blpangle target
This commit adds a new 'blpangle' target that contains libEGL and libGLESv2, so
that we can deploy them as a single versioned module.  However, we still want
to build libEGL and libGLESv2 as separate DLLs so that they can be used by
content_shell.

To do this, we needed to split the 'libEGL' and 'libGLESv2' targets into shared
and static libraries, and change 'libEGL' to a 'loadable_module' so that its
import library is never linked into any projects.

'libGLESv2' cannot be converted to 'loadable_module' because 'libEGL' needs to
link to it.  To prevent its import library from being linked into any other
projects, all other references to it have been removed.  The target has also
been renamed to 'libGLESv2_shared' to prevent future Chromium updates from
adding it as a dependency.

The DllMain from these dlls perform important thread-local-storage setup.  To
use this in a static_library, rename the DllMain functions to 'libEGLMain' and
'libGLESv2Main'.  These functions will be called from blpangle's DllMain.
Also, new dllmain.cpp files have been added to the 'libEGL' and
'libGLESv2_shared' targets.  The new DllMain functions just invoke the
'libEGLMain' and 'libGLESv2Main' functions that are in the static libraries.

The previous main.cpp files also defined an 'error' global function, which have
been renamed to 'libegl_error' and 'libglesv2_error' in order to prevent
potential conflicts in the combined blpangle module.

The code that loads these dlls is in src/ui/gl.  It loads the DLLs and obtains
the eglGetProcAddress function.  However, for blpwtk2, we want it to load the
versioned blpangle DLL instead.  A new 'SetBLPAngleDLLName' function has been
added to src/ui/gl so that we can give it the DLL name.  If the blpangle DLL
name has been set, then src/ui/gl will load the blpangle DLL instead and obtain
the 'blpangle_getProcAddress' function, bypassing libEGL.dll and libGLESv2.dll
entirely.

On the blpwtk2 side, a few things need to be done:
* Create a blpangle target that links 'libEGL_static' and 'libGLESv2_static'.
* DllMain for blpangle needs to call 'libEGLMain' and 'libGLESv2Main'.
* We need to implement a custom get-proc-address procedure.  It needs to be
  able to return the address of all the egl* and gl* functions that are
  exported by 'libEGL' and 'libGLESv2'.  In order to do this, we generate a
  binding file that implements the get-proc-address procedure, using the
  functions defined in the .def files for these dlls.  We use the function
  definitions in src/ui/gl's generate_bindings.py to obtain the function
  signatures.